agents/arm_agent/requirements.txt
google-cloud-aiplatform>=1.42.0
google-adk
sqlalchemy
psycopg2-binary
pandas
numpy
requests
